+++ lib/src/service/tweets/filtering/operator/standalone/context.dart
@@ -1,0 +1,34 @@
+// Copyright 2022 Kato Shinya. All rights reserved.
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided the conditions.
+
+// Project imports:
+import '../operator.dart';
+import '../validation_result.dart';
+
+class Context extends Operator {
+  /// Returns the new instance of [Context].
+  const Context(
+    this.value, {
+    bool negated = false,
+  }) : super(negated);
+
+  factory Context.negated(final String value) => Context(value, negated: true);
+
+  /// The value
+  final String value;
+
+  @override
+  ValidationResult validate() {
+    if (value.isEmpty) {
+      return ValidationResult.failed(
+        'The context must not be an empty string.',
+      );
+    }
+
+    return ValidationResult.succeeded();
+  }
+
+  @override
+  String format() => 'context:$value';
+}